+++ Source/fluffy-obd-lib/diagnostics-engine/elm327/decoders/DecodeFuelRailPressure.hpp
@@ -14,10 +14,10 @@ public:
             throw std::runtime_error("invalid input");
 
         std::string byteA { text[0], text[1] };
-        float valA = Utils::hexToDec(byteA);
+        float valA = static_cast<float>(Utils::hexToDec(byteA));
         std::string byteB { text[3], text[4] };
-        float valB = Utils::hexToDec(byteB);
-        return 0.079 * (256 * valA + valB);
+        float valB = static_cast<float>(Utils::hexToDec(byteB));
+        return 0.079f * (256.0f * valA + valB);
     }
 };
 
@@ -32,9 +32,9 @@ public:
             throw std::runtime_error("invalid input");
 
         std::string byteA { text[0], text[1] };
-        float valA = Utils::hexToDec(byteA);
+        float valA = static_cast<float>(Utils::hexToDec(byteA));
         std::string byteB { text[3], text[4] };
-        float valB = Utils::hexToDec(byteB);
+        float valB = static_cast<float>(Utils::hexToDec(byteB));
         return 10 * (256 * valA + valB);
     }
 };
