+++ Source/fluffy-obd-lib/diagnostics-engine/elm327/decoders/DecodeOxygenSensorCurrent.hpp
@@ -13,10 +13,10 @@ public:
             throw std::runtime_error("invalid input");
 
         std::string byteC { text[0], text[1] };
-        float valC = Utils::hexToDec(byteC);
+        float valC = static_cast<float>(Utils::hexToDec(byteC));
         std::string byteD { text[3], text[4] };
-        float valD = Utils::hexToDec(byteD);
-        return (256.0 * valC + valD) / 256.0 - 128.0;
+        float valD = static_cast<float>(Utils::hexToDec(byteD));
+        return (256.0f * valC + valD) / 256.0f - 128.0f;
     }
 };
