+++ Source/fluffy-obd-lib/diagnostics-engine/elm327/decoders/DecodeTemperature.hpp
@@ -17,7 +17,7 @@ public:
             throw std::runtime_error("invalid input");
 
         std::string byteA { text[0], text[1] };
-        return Utils::hexToDec(byteA) - 40;
+        return static_cast<float>(Utils::hexToDec(byteA) - 40L);
     }
 };
 
@@ -32,10 +32,10 @@ public:
             throw std::runtime_error("invalid input");
 
         std::string byteA { text[0], text[1] };
-        float valA = Utils::hexToDec(byteA);
+        float valA = static_cast<float>(Utils::hexToDec(byteA));
         std::string byteB { text[3], text[4] };
-        float valB = Utils::hexToDec(byteB);
-        return (256 * valA + valB) / 10 - 40;
+        float valB = static_cast<float>(Utils::hexToDec(byteB));
+        return (256.0f * valA + valB) / 10.0f - 40.0f;
     }
 };
